git-svn-id: https://subversion.assembla.com/svn/cmosher@8834 884d43a4-ffbe-47e9-90ca-fd8c37d7eb32

+++ DosMasterToImage/src/dd.java
@@ -105,7 +105,7 @@ public class dd
 		}
 		else if (opt.equals("const"))
 		{
-			constant = Byte.decode(val);
+			constant = Integer.decode(val);
 		}
 		else if (opt.equals("help"))
 		{
